Visual Studio builds: Require Visual Studio 2017
As atkmm master needs C++-14 features that is only fully supported on
Visual Studio 2017, update the projects to become 2017 projects.

Also make sure that we depend on libsigc++-3.x and glibmm-2.52+, and
make them settings in property sheets so that it is easier to bump these
versions if they need to be updated in the future.