generate_defs_and_docs.sh: Update for atk built with meson
When atk is built with meson instead of autotools, generated .h and .c
files are stored in atk/build/atk. Files in that directory shall be read
when atk_docs.xml and the .defs files are generated.

atk_signals.defs contain manually changed lines. Add a .defs.patch file
that automatically makes these changes when atk_signals.defs is regenerated.