build: Add NMake Makefiles for Visual Studio builds
This adds the NMake Makefiles to be used to build atkmm using Visual
Studio 2017 or later, which will supersede the Visual Studio 2017
projects.  This is done to simplify future maintenance, by making use
of the atk/src/filelist.am and atk/atkmm/filelist.am, which is shared
with the autotools build system.