build: Add a set of NMake Makefiles
This is the set of NMake Makefiles that is used to build atkmm using
Visual Studio 2013 and later, which will replace the current Visual
Studio 2013 project files, so that the Visual Studio build files are
easier to maintain.

Note that for the C++-11 version of atkmm, the DLL and library that are
generated from both the visual Studio 2015 and 2017 builds are
atkmm-vc140-1_6.[dll|lib] or atkmm-vc140-d-1_6.[dll|lib] as both Visual
Studio 2015 and 2017 link to the v140 Windows C/C++ runtime DLLs.