Change the ABI to atkmm-2.30
So we can use the 2.28 version numbers for stable releases in the
atkmm-1.6 ABI series.
We don't need to release a stable ABI-parallel atkmm until
we need to release gtkmm 4.0, and that won't happen until GTK+ 4.0.0
happens, and we don't know when that might be.

We've done similar ABI name changes before.