atk/atkmmconfig.h.in: Re-organize ATKMM_API definition
We also activate ATKMM_DLL when we are building with Visual Studio, and
define ATKMM_API as __declspec(dllexport) when ATKMM_DLL and ATKMM_BUILD
are defined.

Since we may be building with sources that are generated by gmmproc that
did not support marking generated items with ATKMM_API, we also leave
around a ATKMM_USE_GENDEF macro that would be defined if we find out
that an older gmmproc was used to generate the sources from the .ccg/.hg
files.