NMake Makefiles: Export using compiler directives if possible
Check the gmmproc-generated sources to see whether they are generated
with a new-enough gmmproc by checking on one of the generated headers,
so that we can tell whether the generated headers have class, function
and method definitions marked with ATKMM_API as needed.

If the headers were generated with a recent enough gmmproc, we define
ATKMM_BUILD and ATKMM_DLL to set ATKMM_API as __declspec(dllexport) to
export the symbols directly during compilation and linking without
building and using gendef.exe, otherwise we fall back to running
gendef.exe on the built object files as we did before.