Visual Studio builds: Clean up build files
glibmm will be updated to be clear of classes that export items making the
built binaries dependent on the exact compiler version and the STL version,
which will eliminate the need to ignore warnings C4251 and C4275.

We will also use the /EHsc compiler flag so that we can also drop the
ignore on warning C4530.